legacy/requirements.txt
click==8.1.3
Flask==2.2.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
psycopg2-binary==2.9
Werkzeug==2.2.2
